Part #2: Refactor run/debug configuration providers
